Address PR review round 8: int power rates, sanitize error messages, fix comment
- Change max_grid_charge_rate/max_pv_charge_rate/min_pv_charge_rate to int in
  InverterConfig — they represent watts and are used with %d formatting in logs
- Sanitize ValidationError in load_config(): use exc.errors() with only field path
  and message (no input_value) to avoid leaking secrets like inverter passwords
- Fix misleading comment in consumption.py: clarify it's an explicit None check
  (not 'or default') so it's clear why 0 is handled correctly as a valid value
- Update test asserting max_grid_charge_rate is float → int

+++ src/batcontrol/config_model.py
@@ -54,9 +54,9 @@ class InverterConfig(BaseModel):
     address: Optional[str] = None
     user: Optional[str] = None
     password: Optional[str] = None
-    max_grid_charge_rate: float = 5000
-    max_pv_charge_rate: float = 0
-    min_pv_charge_rate: float = 0
+    max_grid_charge_rate: int = 5000
+    max_pv_charge_rate: int = 0
+    min_pv_charge_rate: int = 0
     fronius_inverter_id: Optional[str] = None
     fronius_controller_id: Optional[str] = None
     enable_resilient_wrapper: bool = False
